+++ test/confirm/pass_reset_test.exs
@@ -3,11 +3,11 @@ defmodule Phauxth.Confirm.PassResetTest do
   use Plug.Test
   import ExUnit.CaptureLog
 
-  alias Phauxth.{Confirm.PassReset, PhxToken}
+  alias Phauxth.{Confirm.PassReset, TestToken}
 
   setup do
     conn = conn(:get, "/") |> Phauxth.SessionHelper.add_key()
-    valid_email = PhxToken.sign(%{"email" => "froderick@example.com"}, [])
+    valid_email = TestToken.sign(%{"email" => "froderick@example.com"}, [])
     {:ok, %{conn: conn, valid_email: valid_email}}
   end
 
@@ -18,7 +18,7 @@ defmodule Phauxth.Confirm.PassResetTest do
   end
 
   test "reset password fails with expired token" do
-    expired_email = PhxToken.sign(%{"email" => "froderick@example.com"}, [])
+    expired_email = TestToken.sign(%{"email" => "froderick@example.com"}, [])
     params = %{"key" => expired_email, "password" => "password"}
     {:error, message} = PassReset.verify(params, max_age: -1)
     assert message =~ "Invalid credentials"
@@ -26,7 +26,7 @@ defmodule Phauxth.Confirm.PassResetTest do
 
   test "reset fails when confirmed_at is nil" do
     assert capture_log(fn ->
-             valid_key = PhxToken.sign(%{"email" => "brian@example.com"}, [])
+             valid_key = TestToken.sign(%{"email" => "brian@example.com"}, [])
              params = %{"key" => valid_key, "password" => "password"}
              {:error, _} = PassReset.verify(params)
            end) =~ ~s([warn]  user=nil message="unconfirmed user attempting to reset password")
@@ -34,7 +34,7 @@ defmodule Phauxth.Confirm.PassResetTest do
 
   test "reset fails when reset_sent_at is not found" do
     assert capture_log(fn ->
-             valid_key = PhxToken.sign(%{"email" => "igor@example.com"}, [])
+             valid_key = TestToken.sign(%{"email" => "igor@example.com"}, [])
              params = %{"key" => valid_key, "password" => "password"}
              {:error, _} = PassReset.verify(params)
            end) =~ ~s([warn]  user=nil message="no reset token found")